Move template variables extract to utils
